+++ ui/app/src/components/MediaCard/MediaCard.tsx
@@ -109,7 +109,10 @@ function MediaCard(props: MediaCardProps) {
     viewMode = 'card'
   } = props;
   // endregion
-  const { name, path, type } = item;
+  let { name, path, type } = item;
+  if (item.mimeType.includes('audio/')) {
+    type = 'Audio';
+  }
   let iconMap = {
     Page: '@mui/icons-material/InsertDriveFileOutlined',
     Video: '@mui/icons-material/VideocamOutlined',
@@ -118,7 +121,8 @@ function MediaCard(props: MediaCardProps) {
     Component: '@mui/icons-material/ExtensionOutlined',
     Groovy: '@mui/icons-material/CodeRounded',
     JavaScript: '@mui/icons-material/CodeRounded',
-    CSS: '@mui/icons-material/CodeRounded'
+    CSS: '@mui/icons-material/CodeRounded',
+    Audio: '@mui/icons-material/AudiotrackOutlined'
   };
   let icon = { id: iconMap[type] ?? '@mui/icons-material/InsertDriveFileOutlined' };
   const systemIcon = <SystemIcon icon={icon} svgIconProps={{ className: 'media-icon' }} />;