+++ ui/app/src/models/ContentType.ts
@@ -29,9 +29,10 @@ export interface ContentTypeFieldValidation<T = any> {
   level: 'required' | 'suggestion';
 }
 
+// TODO: assess removing `allowedContentTypes` validation in favour of keeping `allowedEmbeddedContentTypes` & `allowedSharedContentTypes` only
 export type ValidationKeys =
   | 'allowedContentTypeTags'
-  | 'allowedContentTypes' // TODO: assess removing this validation in favour of keeping allowedEmbeddedContentTypes & allowedSharedContentTypes only
+  | 'allowedContentTypes'
   | 'allowedEmbeddedContentTypes'
   | 'allowedSharedContentTypes'
   | 'allowedSharedExistingContentTypes'
@@ -57,9 +58,7 @@ export type ValidationKeys =
   | 'allowAudioUpload'
   | 'allowAudioFromRepo';
 
-export type ContentTypeFieldValidations = {
-  [key in ValidationKeys]: ContentTypeFieldValidation;
-};
+export type ContentTypeFieldValidations = Record<ValidationKeys, ContentTypeFieldValidation>;
 
 export interface ValidationResult {
   id: string;