+++ ui/app/src/pages/Global.tsx
@@ -14,12 +14,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import React from 'react';
-import { GlobalAppContextProvider, GlobalApp } from '../components/GlobalApp';
+import { GlobalAppContextProvider, GlobalApp, GlobalAppProps } from '../components/GlobalApp';
 
-export default function Global(topProps: any) {
+export default function Global(props: GlobalAppProps) {
   return (
     <GlobalAppContextProvider>
-      <GlobalApp {...topProps} />
+      <GlobalApp {...props} />
     </GlobalAppContextProvider>
   );
 }