+++ ui/app/src/utils/path.ts
@@ -21,6 +21,7 @@ import LookupTable from '../models/LookupTable';
 import ContentType from '../models/ContentType';
 import { SystemType } from '../models';
 import { v4 as uuid } from 'uuid';
+import { ensureSingleSlash } from './string';
 
 // Originally from ComponentPanel.getPreviewPagePath
 export function getPathFromPreviewURL(previewURL: string): string {
@@ -290,6 +291,7 @@ export function getControllerPath(type: SystemType): string {
   return `/scripts/${type === 'page' ? 'pages' : 'components'}`;
 }
 
+const availableMacrosRegex = /{(objectId|objectGroupId|objectGroupId2|year|month|yyyy|mm|dd|parentPath(\[[0-9]+])?)}/;
 export function processPathMacros(dependencies: {
   path: string;
   objectId: string;
@@ -300,6 +302,15 @@ export function processPathMacros(dependencies: {
   let { path, objectId, objectGroupId, useUUID, fullParentPath } = dependencies;
   let processedPath = path;
 
+  // Remove unrecognized macros.
+  const pathMacros = processedPath.match(/\{([^{}]+)}/g) ?? [];
+  pathMacros.forEach((macro) => {
+    if (!availableMacrosRegex.test(macro)) {
+      processedPath = processedPath.replace(macro, '');
+    }
+  });
+  processedPath = ensureSingleSlash(processedPath);
+
   // The objectGroupId is the first 4 characters of the objectId, so compute if not provided.
   if (objectGroupId === undefined && objectId) {
     objectGroupId = objectId.substring(0, 4);
@@ -309,7 +320,7 @@ export function processPathMacros(dependencies: {
     if (useUUID) {
       processedPath = processedPath.replace('{objectId}', uuid());
     } else {
-      processedPath = path.replace('{objectId}', objectId);
+      processedPath = processedPath.replace('{objectId}', objectId);
     }
   }