+++ ui/guest/rollup.config.mjs
@@ -14,48 +14,26 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import babel from 'rollup-plugin-babel';
-import copy from 'rollup-plugin-copy';
 import commonjs from '@rollup/plugin-commonjs';
 import resolve from '@rollup/plugin-node-resolve';
 import replace from '@rollup/plugin-replace';
-import pkg from './package.json' assert{ type: 'json'};
+import pkg from './package.json' assert { type: 'json' };
+import { swc } from 'rollup-plugin-swc3';
 
-const input = 'src/index.tsx';
-
-const extensions = ['.js', '.jsx', '.ts', '.tsx'];
-
+/** @type {import('rollup').InputPluginOption} */
 const plugins = [
   replace({
     preventAssignment: true,
     'process.env.NODE_ENV': JSON.stringify('production'),
     'process.env.VERSION': JSON.stringify(pkg.version)
   }),
-  babel({
-    exclude: 'node_modules/**',
-    // TODO: @babel/preset-env breaks the build of AMD-style third party libs (e.g. jQuery, js-cookie)
-    presets: ['@babel/preset-react', '@babel/preset-typescript'],
-    plugins: [
-      'babel-plugin-transform-react-remove-prop-types',
-      '@babel/plugin-proposal-nullish-coalescing-operator',
-      '@babel/plugin-proposal-optional-chaining',
-      '@babel/plugin-proposal-class-properties'
-    ],
-    extensions
-  }),
+  swc(),
   resolve({
-    extensions,
+    extensions: ['.js', '.jsx', '.ts', '.tsx'],
     dedupe: ['react', 'react-dom', 'react-is'],
     mainFields: ['module', 'main', 'browser']
   }),
   commonjs({ include: /node_modules|jquery/ }),
-  copy({
-    targets: [
-      { src: 'build/*.umd.js', dest: '../app/public' },
-      { src: 'build/*.umd.js', dest: '../../static-assets/scripts' }
-    ],
-    hook: 'writeBundle'
-  }),
   {
     name: 'rollup-plugin-adjust-studio-ui-imports',
     transform(code, id) {
@@ -73,6 +51,7 @@ const plugins = [
   }
 ];
 
+/** @type {import('rollup').OutputOptions['globals']} */
 const globals = {
   // '@craftercms/content': 'craftercms.content',
   // '@craftercms/search': 'craftercms.search'
@@ -82,46 +61,47 @@ const globals = {
   'prettier/standalone': 'prettierStandalone'
 };
 
+/** @type {import('rollup').RollupOptions['external']} */
 const external = Object.keys(globals);
 
-const baseConfig = {
-  // TODO: Without @babel/preset-env this error doesn't occur.
-  // Addresses rollup's this replaced to undefined
-  context: 'this',
-  onwarn: function (warning, next) {
-    if ( warning.code === 'MODULE_LEVEL_DIRECTIVE' ) return; // you can do this now btw
-    next( warning );
-  },
+/** @type {Partial<import('rollup').RollupOptions>} */
+const commonConfig = {
+  context: 'this', // Addresses Rollup's "this replaced to undefined" warning
+  onwarn(warning, next) {
+    if (warning.code === 'MODULE_LEVEL_DIRECTIVE') return;
+    next(warning);
+  }
 };
 
+/** @type {import('rollup').RollupOptions[]} */
 export default [
   /* UMD build */
   {
-    input,
+    ...commonConfig,
+    input: 'src/index.tsx',
     external,
     plugins,
     output: {
       name: 'craftercms.xb',
-      file: 'build/craftercms-xb.umd.js',
+      file: '../../static-assets/scripts/craftercms-xb.umd.js',
       format: 'umd',
       amd: { id: pkg.craftercms.id },
       globals
-    },
-    ...baseConfig
+    }
   },
 
   /* UMD build for preview landing controller */
   {
+    ...commonConfig,
     input: 'src/preview.ts',
     external,
     plugins,
     output: {
       name: 'craftercms.previewLanding',
-      file: 'build/preview-landing.umd.js',
+      file: '../../static-assets/scripts/preview-landing.umd.js',
       format: 'umd',
       amd: { id: 'craftercms.previewLanding' },
       globals
-    },
-    ...baseConfig
+    }
   }
 ];