+++ ui/uppy/src/components/FileItem/index.jsx
@@ -67,14 +67,17 @@ export default class FileItem extends Component {
             externalMessages={this.props.externalMessages}
           />
           <Buttons
+            uppy={this.props.uppy}
             file={file}
             error={error}
             hideRetryButton={this.props.hideRetryButton}
             showRemoveButton={showRemoveButton}
             validateAndRetry={this.props.validateAndRetry}
+            successfulUploadButton={this.props.successfulUploadButton}
             removeFile={this.props.removeFile}
             retryUpload={this.props.retryUpload}
             i18n={this.props.i18n}
+            validateFilesPolicy={this.props.validateFilesPolicy}
           />
         </div>
         <FileProgress file={file} error={error} isUploaded={isUploaded} />