+++ ui/app/src/CHANGELOG.md
@@ -18,6 +18,7 @@
   * `WorkflowCancellationDialogUI`: prop `items` type changed from an async Resource to sync SandboxItem array.
   * `AuthorFilter` removed, merged with ActivityDashlet
   * `UsersAutocomplete` removed
+  * `UploadDialog`: Added `autoProceed` prop to enable/disable automatic upload after file selection
   * `DropDownMenuButton` component:
     * Internal structure changed so that ListItem wraps the ListItemButton
     * `listItemProps` are routed to the list `ListItem` component instead of the `ListItemButton`
@@ -36,10 +37,16 @@
 * [hooks]
   * `usePreviewUrlControl`: Removed `history` prop. Retrieval of search and navigate (previously called 'push') is now done internally.
 * Removed deprecated `aws-file-upload` and `aws-video` upload controls.
+* Migrated the Studio UI build to Vite/SWC
+* Rollup's XB build to use SWC
+* Upgraded target compilation to ES2022, dropping many code transforms for features that are supported by most modern browsers such as nullish coalescing, optional chaining, object spreading and destructuring.
 
 ## 4.1.5
 * [common-api.js]
   * `CStudioAuthoring.Utils.showConfirmDialog`: Added function overload to receive a `props` style object as first and only argument. The props argument would contain all ConfirmDialog props. Original set of arguments still supported for backward compatibility.
+
+## 4.1.4
+  * `UploadDialog`: Added props `endpoint`, `method`, `headers`, `meta`, `allowedMetaFields`, `useFormData`, `fieldName` and `onFileAdded` for additional control over the upload process.
 
 ## 4.1.3